perbaikan

+++ program_studi/delete.php
@@ -1,5 +1,11 @@
 <?php
+session_start();
 require '../koneksi.php';
+
+if (!isset($_SESSION['login'])) {
+    header("Location: ../login.php");
+    exit;
+}
 
 $id = $_GET['id'];
 mysqli_query($koneksi, "DELETE FROM program_studi WHERE id='$id'");
